Move DocumentViewId into own file, add tests
