+++ p2panda-rs/src/document/document_view_id.rs
@@ -1,0 +1,89 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use std::str::FromStr;
+
+use serde::{Deserialize, Serialize};
+
+use crate::hash::{Hash, HashError};
+use crate::Validate;
+
+/// The identifier of a document view.
+///
+/// Contains the hashes of the document graph tips which is all the information we need to reliably
+/// recreate the document at this certain point in time.
+#[derive(Debug, Eq, PartialEq, Clone, Serialize, Deserialize)]
+pub struct DocumentViewId(Vec<Hash>);
+
+impl DocumentViewId {
+    /// Create a new document view id.
+    pub fn new(graph_tips: Vec<Hash>) -> Self {
+        Self(graph_tips)
+    }
+
+    /// Get the graph tip hashes of this view id.
+    pub fn graph_tips(&self) -> &[Hash] {
+        self.0.as_slice()
+    }
+}
+
+impl Validate for DocumentViewId {
+    type Error = HashError;
+
+    fn validate(&self) -> Result<(), Self::Error> {
+        for hash in &self.0 {
+            hash.validate()?;
+        }
+
+        Ok(())
+    }
+}
+
+/// Convenience method converting a single hash into a document view id.
+///
+/// Converts a `Hash` instance into a `DocumentViewId`, assuming that this document view only
+/// consists of one graph tip hash.
+impl From<Hash> for DocumentViewId {
+    fn from(hash: Hash) -> Self {
+        Self::new(vec![hash])
+    }
+}
+
+/// Convenience method converting a hash string into a document view id.
+///
+/// Converts a hash string into a `DocumentViewId`, assuming that this document view only consists
+/// of one graph tip hash.
+impl FromStr for DocumentViewId {
+    type Err = HashError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(Self::new(vec![Hash::new(s)?]))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use rstest::rstest;
+
+    use crate::hash::Hash;
+    use crate::test_utils::fixtures::random_hash;
+
+    use super::DocumentViewId;
+
+    #[rstest]
+    fn conversion(#[from(random_hash)] hash: Hash) {
+        // Converts any string to `DocumentViewId`
+        let hash_str = "0020d3235c8fe6f58608200851b83cd8482808eb81e4c6b4b17805bba57da9f16e79";
+        let document_id: DocumentViewId = hash_str.parse().unwrap();
+        assert_eq!(
+            document_id,
+            DocumentViewId::new(vec![Hash::new(hash_str).unwrap()])
+        );
+
+        // Converts any `Hash` to `DocumentViewId`
+        let document_id = DocumentViewId::from(hash.clone());
+        assert_eq!(document_id, DocumentViewId::new(vec![hash]));
+
+        // Fails when string is not a hash
+        assert!("This is not a hash".parse::<DocumentViewId>().is_err());
+    }
+}
